refactor: move descriptor methods to dedicated file
