+++ pkg/sif/lookup_test.go
@@ -6,9 +6,6 @@
 package sif
 
 import (
-	"bytes"
-	"io"
-	"path/filepath"
 	"testing"
 )
 
@@ -227,301 +224,6 @@ func TestFromDescr(t *testing.T) {
 	_, _, err = fimg.GetFromDescr(descr)
 	if err == nil {
 		t.Error("fimg.GetFromDescr(descr): should have not found descriptor:", err)
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetData(t *testing.T) {
-	bufferedImage, err := LoadContainer(filepath.Join("testdata", "testcontainer2.sif"), true)
-	if err != nil {
-		t.Fatalf("failed to load container: %v", err)
-	}
-	defer func() {
-		if err := bufferedImage.UnloadContainer(); err != nil {
-			t.Error(err)
-		}
-	}()
-
-	tests := []struct {
-		name string
-		fimg *FileImage
-	}{
-		{"Buffered", &bufferedImage},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			// Get the signature block
-			descr, _, err := tt.fimg.GetFromDescrID(3)
-			if err != nil {
-				t.Fatalf("failed to get descriptor: %v", err)
-			}
-
-			b, err := descr.GetData(tt.fimg)
-			if err != nil {
-				t.Fatalf("failed to get data: %v", err)
-			}
-
-			if got, want := string(b[5:10]), "BEGIN"; got != want {
-				t.Errorf("got data %#v, want %#v", got, want)
-			}
-		})
-	}
-}
-
-func TestGetReader(t *testing.T) {
-	bufferedImage, err := LoadContainer(filepath.Join("testdata", "testcontainer2.sif"), true)
-	if err != nil {
-		t.Fatalf("failed to load container: %v", err)
-	}
-	defer func() {
-		if err := bufferedImage.UnloadContainer(); err != nil {
-			t.Error(err)
-		}
-	}()
-
-	tests := []struct {
-		name string
-		fimg *FileImage
-	}{
-		{"Buffered", &bufferedImage},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			// Get the signature block
-			descr, _, err := tt.fimg.GetFromDescrID(3)
-			if err != nil {
-				t.Fatalf("failed to get descriptor: %v", err)
-			}
-
-			// Read data via Reader and validate data.
-			b := make([]byte, descr.Filelen)
-			if _, err := io.ReadFull(descr.GetReader(tt.fimg), b); err != nil {
-				t.Fatalf("failed to read: %v", err)
-			}
-			if got, want := string(b[5:10]), "BEGIN"; got != want {
-				t.Errorf("got data %#v, want %#v", got, want)
-			}
-		})
-	}
-}
-
-func TestGetName(t *testing.T) {
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	parts, _, err := fimg.GetPartFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetPartFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(parts) != 1 {
-		t.Error("multiple partitions found where expecting 1")
-	}
-
-	if parts[0].GetName() != "busybox.squash" {
-		t.Error(`parts[0].GetName() != "busybox.squash"`)
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetFsType(t *testing.T) {
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	parts, _, err := fimg.GetPartFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetPartFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(parts) != 1 {
-		t.Error("multiple partitions found where expecting 1")
-	}
-
-	fstype, err := parts[0].GetFsType()
-	if err != nil {
-		t.Error("parts[0].GetFsType()", err)
-	}
-
-	if fstype != FsSquash {
-		t.Error("part.GetFsType() should have returned 'FsSquash'")
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetPartType(t *testing.T) {
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	parts, _, err := fimg.GetPartFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetPartFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(parts) != 1 {
-		t.Error("multiple partitions found where expecting 1")
-	}
-
-	parttype, err := parts[0].GetPartType()
-	if err != nil {
-		t.Error("parts[0].GetPartType()", err)
-	}
-
-	if parttype != PartPrimSys {
-		t.Error("part.GetPartType() should have returned 'PartPrimSys'")
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetArch(t *testing.T) {
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	parts, _, err := fimg.GetPartFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetPartFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(parts) != 1 {
-		t.Error("multiple partitions found where expecting 1")
-	}
-
-	arch, err := parts[0].GetArch()
-	if err != nil {
-		t.Error("parts[0].GetArch()", err)
-	}
-
-	if trimZeroBytes(arch[:]) != HdrArchAMD64 {
-		t.Logf("|%s|%s|\n", arch[:hdrArchLen-1], HdrArchAMD64)
-		t.Error("part.GetArch() should have returned 'HdrArchAMD64':", err)
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetHashType(t *testing.T) {
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	sigs, _, err := fimg.GetSignFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetSignFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(sigs) != 1 {
-		t.Error("multiple signatures found where expecting 1")
-	}
-
-	hashtype, err := sigs[0].GetHashType()
-	if err != nil {
-		t.Error("sigs[0].GetHashType()", err)
-	}
-
-	if hashtype != HashSHA384 {
-		t.Error("sig.GetHashType() should have returned 'HashSHA384'")
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetEntity(t *testing.T) {
-	expected := []byte{159, 43, 108, 54, 217, 153, 163, 233, 28, 179, 16, 71, 32, 103, 21, 144, 193, 45, 66, 34}
-
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	sigs, _, err := fimg.GetSignFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetSignFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(sigs) != 1 {
-		t.Error("multiple signatures found where expecting 1")
-	}
-
-	entity, err := sigs[0].GetEntity()
-	if err != nil {
-		t.Error("sigs[0].GetEntity()", err)
-	}
-
-	if bytes.Equal(expected, entity[:len(expected)]) == false {
-		t.Error("sig.GetEntity(): didn't get the expected entity, got:", entity[:len(expected)])
-	}
-
-	// unload the test container
-	if err = fimg.UnloadContainer(); err != nil {
-		t.Error("UnloadContainer(fimg):", err)
-	}
-}
-
-func TestGetEntityString(t *testing.T) {
-	expected := "9F2B6C36D999A3E91CB3104720671590C12D4222"
-
-	// load the test container
-	fimg, err := LoadContainer("testdata/testcontainer2.sif", true)
-	if err != nil {
-		t.Error("LoadContainer(testdata/testcontainer2.sif, true):", err)
-	}
-
-	sigs, _, err := fimg.GetSignFromGroup(DescrDefaultGroup)
-	if err != nil {
-		t.Error("fimg.GetSignFromGroup(DescrDefaultGroup): should have found descriptor:", err)
-	}
-
-	if len(sigs) != 1 {
-		t.Error("multiple signatures found where expecting 1")
-	}
-
-	entity, err := sigs[0].GetEntityString()
-	if err != nil {
-		t.Error("sigs[0].GetEntityString()", err)
-	}
-
-	if expected != entity {
-		t.Error("sig.GetEntityString(): didn't get the expected entity")
 	}
 
 	// unload the test container
